Time: 115 ms (48.70%), Space: 13.4 MB (37.42%) - LeetHub

+++ 18-4sum/18-4sum.cpp
@@ -22,7 +22,7 @@ public:
 
                         while(low<high)
                         {
-                            if(long(nums[low] + nums[high]) == sum)
+                            if(nums[low] + nums[high] == sum)
                             {
                                 vector<int> arr;
                                 arr.push_back(nums[i]);
